\name{BANDITS-package}
\alias{BANDITS-package}
\alias{BANDITS}
\docType{package}
\title{
\packageTitle{BANDITS}
}
\description{
\packageDescription{BANDITS}

Questions relative to BANDITS should be either written to the \href{https://support.bioconductor.org}{Bioconductor support site}, tagging the question with 'BANDITS', or reported as a new issue at \href{https://github.com/SimoneTiberi/BANDITS/issues}{BugReports}.

To access the vignettes, type: \code{browseVignettes("BANDITS")}.
}
\author{
\packageAuthor{BANDITS}

Maintainer: \packageMaintainer{BANDITS}
}
\keyword{
differential splicing, differential transcript usage, DTU, Dirichlet-multinomial, Bayesian hierarchical modelling, data augmentation
}