% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_bandits.R
\name{binomial_bandit_sim}
\alias{binomial_bandit_sim}
\title{Estimate the Bayesian posterior probability of each alternative
being the best binomial bandit at time t (simulation version)}
\usage{
binomial_bandit_sim(x, n, .alpha = 1, .beta = 1, ndraws = 5000, ncp = 0)
}
\arguments{
\item{x}{number of success}

\item{n}{number of trials}

\item{.alpha}{alpha of beta distribution}

\item{.beta}{beta of beta distribution}

\item{ncp}{prior center}
}
\value{
optimal probabilities of each arm w_{a, t}
}
\description{
Estimate the Bayesian posterior probability of each alternative
being the best binomial bandit at time t (simulation version)
}
\seealso{
Eq (11)
}

