% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{experiment}
\alias{experiment}
\title{experiment}
\usage{
experiment(simple_learner, bandit, number_of_runs, save_logs = F)
}
\description{
takes a multiarm bandit and a learner and runs through the training process.
}
\author{
Joe Peskett
}
