% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_data.R
\name{input_data}
\alias{input_data}
\title{A \code{\linkS4class{BANDITS_data}} object, generated with \code{\link{create_data}}}
\arguments{
\item{input_data}{a \code{\linkS4class{BANDITS_data}} object.}
}
\description{
A \code{\linkS4class{BANDITS_data}} object, generated with \code{\link{create_data}}
}
\examples{
# Computed as shown in the vignettes, see: browseVignettes("BANDITS")

# create data and filter internally lowly abundant transcripts:
#input_data = create_data(gene_to_transcript = gene_tr_id,
#                           path_to_eq_classes = equiv_classes_files, eff_len = eff_len, 
#                           n_cores = 2,
#                           transcripts_to_keep = transcripts_to_keep)

# load the pre-computed data:
data("input_data", package = "BANDITS")
input_data

}
\seealso{
\code{\link{create_data}}, \code{\linkS4class{BANDITS_data}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
