% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prec.R
\name{prec}
\alias{prec}
\title{Estimates for the log-precision parameter}
\arguments{
\item{prec}{a \code{list} with two elements:
\itemize{
  \item 'prior', a numeric \code{vector} containing the mean and standard deviation prior estimates
  for the log-precision parameter;
  \item 'genewise_log_precision', a numeric \code{vector} containing the gene-wise precision estimates
  of the log-precision parameter.
}}
}
\description{
A list containing the estimates for the log-precision parameter of the Dirichlet-multinomial distribution.
}
\examples{
# load the precision estimates
data(prec, package = "BANDITS")

# mean and standard deviation estimates of the log-precision parameter:
prec$prior

# gene-wise estimates of the log-precision parameter:
prec$genewise_log_precision

# Plot the histogram of the genewise log-precision estimates.
# The black solid line represents the normally distributed prior distribution 
# for the log-precision parameter.
plot_precision(prec)

}
\seealso{
\code{\link{prior_precision}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
