% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_results.R
\name{results}
\alias{results}
\title{Results of the DTU test, generated with \code{\link{test_DTU}}}
\arguments{
\item{results}{a \code{\linkS4class{BANDITS_test}} object containing the results of the DTU test,
generated with \code{\link{test_DTU}}.}
}
\description{
Results of the DTU test, generated with \code{\link{test_DTU}}
}
\examples{
# Object 'results' is generated via 'test_DTU' as shown in the vignettes: 
# see browseVignettes("BANDITS").

# load the pre-computed results:
data("results", package = "BANDITS")
results

}
\seealso{
\code{\link{test_DTU}}, \code{\linkS4class{BANDITS_test}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
