% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{simple_learner}
\alias{simple_learner}
\title{simple_learner}
\usage{
simple_learner(n_bandit, exploration)
}
\arguments{
\item{n_bandit}{a number to pass to the multiarm bandit function}

\item{exploration}{the probability of a random value being selected}

\item{bandit}{the function to create the multiarm bandit}
}
\description{
function to return a learner given the problem passed to it and learning implementation
}
\author{
Joe Peskett
}
