% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{value_update}
\alias{value_update}
\title{value_update}
\usage{
value_update(OldEst, Target, Step_size)
}
\description{
Update value of action
NewEst = OldEst + step(Target-OldEst)
}
\author{
Joe Peskett
}
