% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gene_tr_id.R
\name{gene_tr_id}
\alias{gene_tr_id}
\title{Gene-transcript matching}
\arguments{
\item{gene_tr_id, }{a \code{data.frame} containing the matching between gene (1st column) and transcript identifiers (2nd column).}
}
\description{
A list containing the estimates for the log-precision parameter of the Dirichlet-multinomial distribution.
}
\examples{
# load the Gene-Transcript data.frame and visualize its top
data(gene_tr_id, package = "BANDITS")
head(gene_tr_id)

}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
