% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_data.R
\name{input_data}
\alias{input_data}
\title{A \code{\linkS4class{BANDITS_data}} object, generated with \code{\link{create_data}}}
\arguments{
\item{input_data}{a \code{\linkS4class{BANDITS_data}} object, generated via \code{\link{create_data}}.}
}
\description{
A \code{\linkS4class{BANDITS_data}} object, generated with \code{\link{create_data}}
}
\examples{
# Object 'input_data' is generated via 'create_data' as shown in the vignettes:
# see browseVignettes("BANDITS").

# load the pre-computed data:
data("input_data", package = "BANDITS")
input_data

}
\seealso{
\code{\link{create_data}}, \code{\linkS4class{BANDITS_data}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
