% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_precision.R
\name{plot_precision}
\alias{plot_precision}
\title{Plot the log-precision estimates}
\usage{
plot_precision(prior)
}
\arguments{
\item{prior}{the prior of the log-precision parameter, computed via \code{\link{prior_precision}}.}
}
\value{
A plot.
}
\description{
Plot the log-precision estimates
}
\examples{
## Preliminary information

# specify the directory of the internal data:
data_dir = system.file("extdata", package = "BANDITS")
data_dir

# load gene_to_transcript matching:
data("gene_tr_id", package = "BANDITS")
# gene_tr_id contains transcripts ids on the first column
# and the corresponding gene ids on the second column:
head(gene_tr_id)

# Specify the directory of the transcript level estimated counts.
sample_names = paste0("sample", seq_len(4))
quant_files = file.path(data_dir, sample_names, "quant.sf")
file.exists(quant_files)

# Load the transcript level estimated counts via tximport:
library(tximport)
txi = tximport(files = quant_files, type = "salmon", txOut = TRUE)
counts = txi$counts
head(counts)



## Optional (recommended): infer an informative prior for the precision parameter

# Use the same filtering criteria as in \\code{\\link{filter_transcripts}}; 
# if transcript pre-filtering is not performed, set \\code{min_transcript_proportion},
# \\code{min_transcript_counts} and \\code{min_gene_counts} to 0.

#set.seed(61217)
#prec = prior_precision(gene_to_transcript = gene_tr_id, transcript_counts = counts,
#                       min_transcript_proportion = 0.01, min_transcript_counts = 10,
#                       min_gene_counts = 20, n_cores = 2)

# load the pre-computes precision estimates:
data(prec, package = "BANDITS")

# Plot the histogram of the genewise log-precision estimates.
# The black solid line represents the normally distributed prior distribution 
# for the log-precision parameter.
plot_precision(prec)

}
\seealso{
\code{\link{test_DTU}}, \code{\link{prior_precision}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
